Renamed: component WidgetPaginator -> SelectCurrentPage
