+++ apps/frontend/src/components/SelectCurrentPage.vue
